Refactored package dateinfer in terms of architecture and code style
